server/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
Flask-PyMongo>=2.3
pymongo[srv]>=4.0
bcrypt==5.0.0
flask-jwt-extended
python-dotenv==1.1.1
PyJWT==2.10.1
Werkzeug==3.1.3